+++ 2019/src/main/java/com/mdiazv/advent/Day5.java
@@ -26,11 +26,11 @@ public class Day5 implements Day {
 	private InputStream getInputStream() {
 		return this.getClass().getResourceAsStream("/input/5.txt");
 	}
-	public static int[] parseProgram(InputStream in) { 
+	public static long[] parseProgram(InputStream in) {
 		try {
 			String text = new BufferedReader(new InputStreamReader(in)).readLine();
 			return Stream.of(text.split(","))
-				.mapToInt(Integer::valueOf)
+				.mapToLong(Long::valueOf)
 				.toArray();
 		} catch (java.io.IOException e) {
 			System.err.println("Could not load program");
@@ -39,11 +39,11 @@ public class Day5 implements Day {
 		}
 	}
 	public void run() {
-		int[] program = parseProgram(getInputStream());
+		long[] program = parseProgram(getInputStream());
 		IntCodeComputer computer = new IntCodeComputer(program);
-		Vector<Integer> output = computer.executeWith(new int[]{1});
+		Vector<Long> output = computer.executeWith(new long[]{1});
 		System.out.println("Diagnostics output for ID 1: "+ output);
-		output = computer.executeWith(new int[]{5});
+		output = computer.executeWith(new long[]{5});
 		System.out.println("Diagnostics output for ID 5: "+ output);
 	}
 }
